+++ backend/tests/userServices.test.js
@@ -1,5 +1,5 @@
 const mongoose = require("mongoose");
-const UserSchema = require("../models/users");
+const UserSchema = require("../models/userSchema");
 const userServices = require("../routes/userServices");
 const db = require("../db.js");
 const { MongoMemoryServer } = require("mongodb-memory-server");
